Use uint32_t for a variable that's passed to a function expecting a uint32_t.

+++ src/ast_vgatool.c
@@ -2709,7 +2709,7 @@ void vGetDefaultSettings(ScrnInfoPtr pScrn)
 Bool InitVGA(ScrnInfoPtr pScrn, ULONG Flags)
 {
    ASTRecPtr pAST;
-   ULONG ulData;
+   uint32_t ulData;
 
    pAST = ASTPTR(pScrn);
 
@@ -3005,7 +3005,7 @@ GetVGAEDID(ScrnInfoPtr pScrn, unsigned char *pEDIDBuffer)
 Bool bInitAST1180(ScrnInfoPtr pScrn)
 {
     ASTRecPtr pAST;	
-    ULONG ulData;
+    uint32_t ulData;
 
     pAST = ASTPTR(pScrn);
